+++ CHANGELOG.md
@@ -5,6 +5,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Add Windows implementation (see [GH-1]).
+
+[Gh-1]: https://github.com/lunaryorn/gethostname.rs/pull/1
+
 ## 0.1.0 – 2019-01-20
 Initial release.